### 2024-02-23 #6397 @kyleniemeyer
@jgostick thanks for raising the concern - I think it looks small because that info was just from the `joss-paper` branch, vs. the main branch which has more.
