### 2024-03-08 #6397 @jgostick
Hey @ericfell, and ideas for people to review this?  There is a serious lack of electrochemistry expertise in the JOSS reviewer database.  2 of the 5 of them are my former students (Amin and Tom) so I hesitate to ask them for COI reasons.
