### 2024-03-08 #6397 @ericfell
@jgostick I feel you, I'm hoping to be a reviewer myself once I've had more experience with JOSS.
I'm assuming the requirement is someone with both sufficient expertise in the field AND capable in the coding language of the submission? Plenty of echem matlab users out there but for python I'd suggest Johan Hjelm at Technical University of Denmark. I know of others but all would be COI. Is there an official _recommend a reviewer_ process for JOSS?
