### 2024-03-25 #6537 @editorialbot
[REVIEW]: RFBzero: A Python package for zero-dimensional simulation of redox flow battery cycling
**Submitting author:** @ericfell (<a href="http://orcid.org/0000-0003-2046-1480">Eric M. Fell</a>)
**Repository:** https://github.com/ericfell/rfbzero
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** 1.0.1
