### 2024-04-29 ericfell/rfbzero#35 @MaximevdHeijden
[JOSS Review] – Community guidelines 
Dear authors,

To improve the community guidelines, could you please make a 'Contribution' section to improve the visibility for third parties? Now it is mentioned before the installation, which is easy to overlook when looking for the contribution guidelines.

This issue is related to openjournals/joss-reviews#6537
