### 2024-04-29 ericfell/rfbzero#36 @MaximevdHeijden
[JOSS Review] – Software paper
Dear authors,

Could you improve the state of the field? In the Background section it is not clear enough what the new features are of this software package versus existing packages, especially in comparison with the cited references, i.e., which features are the other packages missing?

This issue is related to openjournals/joss-reviews#6537
