### 2024-05-30 ericfell/rfbzero#37 @MaximevdHeijden
[JOSS Review] - Suggestions for the documentation
Dear authors,

Some very minor suggestions for the documentation:

- Could you add the installation section to the documentation? Copy-paste from your README file is more than fine, but then everything can be found in the same documentation.

- The crossover documentation mentions that the class is still limited to symmetric cells and that current-driven crossover cannot be simulated. If you are not working on this yourself anymore by any chance, it would be nice to make a list of possible extensions of the code for others to contribute to.

This issue is related to https://github.com/openjournals/joss-reviews/issues/6537
