### 2024-05-30 #6537 @MaximevdHeijden
Dear authors,
First of all, my apologies for the delay.
The software package looks great! The installation works well, the documentation is clear and complete, the examples are easy to follow, and the code is user friendly and easy to adjust for your own simulations. Besides the minor comments in the three issues that I opened I don’t see any other adjustments that need to be made to the software/documentation/paper.
