### 2024-05-31 ericfell/rfbzero#35 @hjelmj
Dear authors, I agree with @MaximevdHeijden, but I also think it could be useful to add it to the repository as a CONTRIBUTING.md file where the contribution instructions are provided.
