### 2024-06-04 5d20651
add CONTRIBUTING, re joss review #35
