### 2024-06-04 198d8d7
add contribution guidlines to docs FAQ, re joss-review #35
