+++ docs/source/faq.rst
@@ -1,2 +1,25 @@
 Frequently Asked Questions
 ==========================
+
+How can I contribute to rfbzero.py?
+-------------------------------------
+First off, thank you for taking the time to contribute!
+Everyone is welcome to contribute and all types of contributions are encouraged and valued. This includes code reviews, bug fixes, new features, examples, documentation, community participation, etc.
+
+Feature requests and bug reports
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+If you want to make a suggestion for a new feature, please `make an
+issue <https://github.com/ericfell/rfbzero/issues/new/choose>`_
+including as much detail as possible.
+
+Contributing code
+~~~~~~~~~~~~~~~~~
+
+The preferred method for contributing code to rfbzero.py is to fork
+the repository on GitHub and submit a "pull request" (PR).
+More detailed information on how to get started developing rfbzero.py
+can be found in
+`CONTRIBUTING.md <https://github.com/ericfell/rfbzero/blob/main/CONTRIBUTING.md>`_.
+
+Feel free to reach out via GitHub issues with any questions!
