### 2024-06-04 2d2c2c4
add installation info in examples notebook, re joss-review #37
