### 2024-06-05 11edab6
fix minor wording, re comments on #39
