### 2024-06-05 9105cc2
fix small typos, re comments #40
