+++ docs/source/faq.rst
@@ -9,7 +9,7 @@ Everyone is welcome to contribute and all types of contributions are encouraged 
 Feature requests and bug reports
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-If you want to make a suggestion for a new feature, please `make an
+If you want to make a suggestion for a new feature, please `create an
 issue <https://github.com/ericfell/rfbzero/issues/new/choose>`_
 including as much detail as possible.
 
@@ -17,7 +17,7 @@ Contributing code
 ~~~~~~~~~~~~~~~~~
 
 The preferred method for contributing code to rfbzero.py is to fork
-the repository on GitHub and submit a "pull request" (PR).
+the repository on GitHub and create a pull request with the proposed changes.
 More detailed information on how to get started developing rfbzero.py
 can be found in
 `CONTRIBUTING.md <https://github.com/ericfell/rfbzero/blob/main/CONTRIBUTING.md>`_.
