### 2024-06-05 ericfell/rfbzero#35 @ericfell
Thank you both for your suggestions.
We've added a contribution section to the documents FAQ and a CONTRIBUTING.md in the root repo,
via #40 (now merged).
related to openjournals/joss-reviews#6537
