### 2024-06-05 ericfell/rfbzero#41 @ericfell
Current-driven crossover
## Description

In some cases we may want to model current-driven crossover of redox-actives, rather than purely concentration-gradient-driven crossover (as is already included in the `Crossover` class).

## Motivation

Current-driven crossover is not a dominant effect for the large redox-active organic molecules with which `rfbzero` was primarily built to model, but is definitely important to include when simulating the metal-ion RFB systems (Fe, V, Cr, etc.), or systems employing small molecules and size-exclusion membranes.

## Possible Implementation

If we can assume dilute solution theory is applicable, then the equations for current-driven crossover have previously been adapted for zero-dimensional cases, see [The Influence of Electric Field on Crossover in Redox-Flow Batteries](https://iopscience.iop.org/article/10.1149/2.0031601jes) and later work, [A Method for Quantifying Crossover in Redox Flow Cells through Compositionally Unbalanced Symmetric Cell Cycling](https://iopscience.iop.org/article/10.1149/1945-7111/ace938). This should be straightforward to implement in `Crossover`, but more input parameters will be needed to describe membrane properties (membrane conductivity, electro-osmotic coefficient, concentration of fixed ion sites, etc.).

## Additional Context

*No response*
