### 2024-06-05 ericfell/rfbzero#37 @ericfell
Thank you for the suggestions.

Regarding the first bullet point, we've added the installation section to the documentation [examples notebook](https://github.com/ericfell/rfbzero/blob/main/docs/source/examples/simulate_RFB_example.ipynb) in #40, and the same information can be found on the homepage [readthedocs for rfbzero](https://rfbzero.readthedocs.io/en/latest/index.html).

Regarding the second point, we've opened issue #41 as a feature request to include current-driven crossover. Guiding information for this feature is provided there.

related to openjournals/joss-reviews#6537
