### 2024-06-11 1ae187e
Print CyclingStatus value to fix str/Enum behavior change for CPython 3.11+
