### 2024-06-11 8d2fa59
Update GitHub actions to use multiple Python versions for tests
