### 2024-06-11 ericfell/rfbzero#38 @jeremyfell
Thank you for raising this! This was due to a [CPython `3.11+` issue](https://github.com/python/cpython/issues/100458) where the behavior of str/Enum mixins was changed.
We have GitHub actions that run the tests on each pull request, but previously only Python `3.10` was used so this went unnoticed.

This is now fixed in https://github.com/ericfell/rfbzero/pull/42, and the GitHub actions have been updated to run the tests with multiple Python versions.

Related to https://github.com/openjournals/joss-reviews/issues/6537.
