### 2024-06-12 #6537 @ericfell
Thanks @jgostick all checks look good.
version for JOSS paper: v1.0.1
Zenodo DOI [10.5281/zenodo.11594955](https://zenodo.org/doi/10.5281/zenodo.11594955)
