### 2024-06-20 #6537 @jgostick
Hi @ericfell...the paper looks good, but the capitalization of the paper titles in the bibliography is not consistent.  Normal sentence case is preferred.
