### 2024-06-20 aa52e0e
make citation title normal sentence case
