+++ joss-paper/paper.bib
@@ -1,7 +1,7 @@
 @Article{2018_Konig_JPOWERSOURCE,
   author   = {König, S. and Suriyah, M. R. and Leibfried, T.},
   journal  = {Journal of Power Sources},
-  title    = {Validating and improving a zero-dimensional stack voltage model of the {Vanadium} {Redox} {Flow} {Battery}},
+  title    = {Validating and improving a zero-dimensional stack voltage model of the vanadium redox flow battery},
   year     = {2018},
   issn     = {0378-7753},
   month    = feb,
@@ -15,7 +15,7 @@
 @Article{2018_Murthy_JES,
   author    = {Sri Krishna Murthy and Ashwini Kumar Sharma and Clement Choo and Erik Birgersson},
   journal   = {Journal of The Electrochemical Society},
-  title     = {Analysis of {C}oncentration {O}verpotential in an {A}ll-{V}anadium {R}edox {F}low {B}attery},
+  title     = {Analysis of concentration overpotential in an all-vanadium redox flow battery},
   year      = {2018},
   number    = {9},
   pages     = {A1746--A1752},
@@ -42,7 +42,7 @@
 @Article{2019_Lee_JECHEMENERGCONVSTOR,
   author     = {Lee, Seong Beom and Mitra, Kishalay and Pratt, III, Harry D. and Anderson, Travis M. and Ramadesigan, Venkatasailanathan and Chalamala, Babu R. and Subramanian, Venkat R.},
   journal    = {Journal of Electrochemical Energy Conversion and Storage},
-  title      = {Open {Data}, {Models}, and {Codes} for {Vanadium} {Redox} {Batch} {Cell} {Systems}: {A} {Systems} {Approach} {Using} {Zero}-{Dimensional} {Models}},
+  title      = {Open data, models, and codes for vanadium redox batch cell systems: {A} systems approach using zero-dimensional models},
   year       = {2019},
   issn       = {2381-6872},
   month      = sep,
@@ -56,7 +56,7 @@
 @Article{2020_Kwabi_CHEMREV,
   author     = {Kwabi, David G. and Ji, Yunlong and Aziz, Michael J.},
   journal    = {Chemical Reviews},
-  title      = {Electrolyte {Lifetime} in {Aqueous} {Organic} {Redox} {Flow} {Batteries}: {A} {Critical} {Review}},
+  title      = {Electrolyte lifetime in aqueous organic redox flow batteries: {A} critical review},
   year       = {2020},
   issn       = {0009-2665},
   month      = jul,
@@ -72,7 +72,7 @@
 @Article{2021_Modak_JES,
   author     = {Sanat Modak and David G. Kwabi},
   journal    = {Journal of The Electrochemical Society},
-  title      = {{A} {Z}ero-{D}imensional {M}odel for {E}lectrochemical {B}ehavior and {C}apacity {R}etention in {O}rganic {F}low {C}ells},
+  title      = {{A} zero-dimensional model for electrochemical behavior and capacity retention in organic flow cells},
   year       = {2021},
   month      = {aug},
   number     = {8},
@@ -85,7 +85,7 @@
 @Article{2022_Neyhouse_JES,
   author    = {Neyhouse, Bertrand J. and Lee, Jonathan and Brushett, Fikile R.},
   journal   = {Journal of The Electrochemical Society},
-  title     = {Connecting {Material} {Properties} and {Redox} {Flow} {Cell} {Cycling} {Performance} through {Zero}-{Dimensional} {Models}},
+  title     = {Connecting material properties and redox flow cell cycling performance through zero-dimensional models},
   year      = {2022},
   issn      = {1945-7111},
   month     = sep,
@@ -101,7 +101,7 @@
[…]
