### 2024-06-20 #6537 @ericfell
thanks @jgostick I've made them all normal sentence case now (pushed to joss-paper branch)
